added a FakeStringTranslator that fakes the translation of text.
